+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Added
+
+* Render multi-line strings as block-scalars in Commodore's YAML dump helpers ([#294])
+
 ## [v0.5.0] - 2021-03-12
 
 ### Added
@@ -344,3 +348,4 @@ Initial implementation
 [#287]: https://github.com/projectsyn/commodore/pull/287
 [#289]: https://github.com/projectsyn/commodore/pull/289
 [#290]: https://github.com/projectsyn/commodore/pull/290
+[#294]: https://github.com/projectsyn/commodore/pull/294